tidied deployment stuff
